Remove mixins
They don't really have a place anymore in an Octane world with native classes, and were kind of redundant anyway.
